You shall not use null in Scala. Part I.
It's ok for Python to use it's equivalent, None, though.
A follow up commit will follow where we rely more on implicit arguments
as it was initially intended, I think.

Every call from Python to Scala needs to translate None to scala.None
before executing it.
There are more nulls flying around so this is only part one.