Part II of getting rid off nulls - CSVAttributes
CSVAttributes are often optional and depending on the context,
also has different defaults (if coming from Python or not).

It's ok for Python to use null's equivalent, None, though.
A follow up commit will follow where we rely more on implicit arguments
as it was initially intended, I think.

Every call from Python to Scala needs to translate None to scala.None
before executing it.