Part XIII of nulls - SchemaDiscoveryHelper
Got rid off the ugly null initialization for a list and used a
proper Map having columns as keys. The schema entries have to be
ordered in the same order as the columns of the RDD, so one cannot use a
simple map method, but rather create a tuple List manually (see `discoverySchema`).

Also got rid of some ugly imperative code in canConvertToyyy....